backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
PyMuPDF
unstructured
pytesseract
pillow
sentence-transformers
transformers
torch
chromadb
litellm
ollama
google-generativeai
spacy
scikit-learn
presidio-analyzer
presidio-anonymizer
langfuse
promptfoo
python-dotenv
httpx
aiofiles
fuzzywuzzy
python-Levenshtein
reportlab